refactor: remove registerPage from components/pages
